requirements.txt
torch==1.7.1
torchvision==0.8.2
pandas
numpy
gensim==4.1.2
Pillow==8.4.0
tqdm
tensorboard
wandb